Setup necessary class variables when inheriting from Workflow
Not sure if the early return from variables_setup if already setup is necessary. Seems like probably not. If you define a workflow that inherits from another workflow that inherits from the base workflow you'll hit the 'inherited' hook twice... But since these are class instance variables not class variables (@state not @@state), the variables aren't shared with child classes, so it's not like the child class is going to clobber parent values anyway (since it won't get any parent values, there's nothing to clobber). Even if it's unnecessary, it sort of makes it clear that nothing is going to get initialized twice.