Remove comment about clarifying implicit dependency on Workflow mixins
Yes, it's a bummer that the dependency is implicit. However, with the initialization of the variables (e.g. @states) into the base Workflow class, I think it's easier to see that states is supposed to be shared. And the naming of the modules (prefixing with Workflow, e.g. in WorkflowValidation) I think suggests that it's intended to be used in that module. With those two things I don't think it's very hard to figure out where the self.states is coming from. I think having the various pieces separated into modules makes it easier to reason about the different sets of functionality included in a workflow (as opposed to having it in one giant file), and so the trade off is worth it. I'm sure there are other approaches, so if a better one comes up great, but this is fine for now