Update workflow transition options note
the work_queue one might be the hardest to do... i'm not totally positive it's necessary... but say you had a state transition that required a ton of memory or a beefy gpu or something. If every worker is pulling from the same work queue, and all work in a workflow is in that same work queue... how would you accomplish this?

if state transitions can have their own work queue, then you could spin up worker pods and have them work on just that queue. likewise other pods wouldn't listen to the queue. It's unknown to me how exactly that would work... but I could imagine the Engine.run optionally taking some config that specifies what work queues to listen to. By default, it just assumes there's only a single work queue per workflow and just listens on that. Maybe it could even detect if there are more than 1 work queues in a workflow, and blow up on start up if you didn't explicitly tell it which queues to run

and the condition idea? i scrapped it. adds complexity for not much value because the user can easily handle conditions themselves in the process function